requirements.txt
opencv_contrib_python==3.4.2.16
torchvision==0.5.0
matplotlib==3.0.1
Shapely==1.7.1
cityscapesScripts==2.1.7
mock==4.0.3
cloudpickle==0.5.3
numpy==1.16.2
torchfile==0.1.0
ipdb==0.12.2
imagesize==1.2.0
scikit_image==0.16.2
scipy==1.4.1
tabulate==0.8.7
tqdm==4.32.1
termcolor==1.1.0
torch==1.4.0
sphinx_rtd_theme==0.5.0
fvcore==0.1
numexpr==2.7.1
psutil==5.7.0
recommonmark==0.6.0
Pillow==8.0.1
scikit_learn==0.23.2
